refactor download route
