+++ apps/app/src/server/routes/attachment/index.ts
@@ -1,2 +1,3 @@
 export * from './get';
 export * from './get-brand-logo';
+export * from './download';
